restructured exceptions into separate file, updated respective imports.
