+++ lasif/components/adjoint_sources.py
@@ -7,8 +7,8 @@ import os
 import numpy as np
 from lasif.utils import process_two_files_without_parallel_output
 
-from lasif import LASIFNotFoundError
-from .component import Component
+from lasif.exceptions import LASIFNotFoundError
+from lasif.component import Component
 from lasif.tools.adjoint.adjoint_source import calculate_adjoint_source
 
 # Map the adjoint source type names to functions implementing them.
